+++ src/PimEnterprise/Bundle/CatalogRuleBundle/Engine/ProductRuleApplier/ProductsUpdater.php
@@ -14,8 +14,8 @@ namespace PimEnterprise\Bundle\CatalogRuleBundle\Engine\ProductRuleApplier;
 use Akeneo\Bundle\RuleEngineBundle\Model\RuleInterface;
 use Doctrine\Common\Util\ClassUtils;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdaterInterface;
 use Pim\Bundle\CatalogBundle\Updater\ProductTemplateUpdaterInterface;
-use Pim\Bundle\CatalogBundle\Updater\ProductUpdaterInterface;
 use PimEnterprise\Bundle\CatalogRuleBundle\Model\ProductCopyValueActionInterface;
 use PimEnterprise\Bundle\CatalogRuleBundle\Model\ProductSetValueActionInterface;
 
@@ -26,21 +26,21 @@ use PimEnterprise\Bundle\CatalogRuleBundle\Model\ProductSetValueActionInterface;
  */
 class ProductsUpdater
 {
-    /** @var ProductUpdaterInterface */
-    protected $productUpdater;
+    /** @var ProductFieldUpdaterInterface */
+    protected $productFieldUpdater;
 
     /** @var ProductTemplateUpdaterInterface */
     protected $templateUpdater;
 
     /**
-     * @param ProductUpdaterInterface         $productUpdater
+     * @param ProductFieldUpdaterInterface    $productFieldUpdater
      * @param ProductTemplateUpdaterInterface $templateUpdater
      */
     public function __construct(
-        ProductUpdaterInterface $productUpdater,
+        ProductFieldUpdaterInterface $productFieldUpdater,
         ProductTemplateUpdaterInterface $templateUpdater
     ) {
-        $this->productUpdater  = $productUpdater;
+        $this->productFieldUpdater  = $productFieldUpdater;
         $this->templateUpdater = $templateUpdater;
     }
 
@@ -99,7 +99,7 @@ class ProductsUpdater
     protected function applyCopyAction(array $products, ProductCopyValueActionInterface $action)
     {
         foreach ($products as $product) {
-            $this->productUpdater->copyData(
+            $this->productFieldUpdater->copyData(
                 $product,
                 $product,
                 $action->getFromField(),
@@ -127,7 +127,7 @@ class ProductsUpdater
     protected function applySetAction(array $products, ProductSetValueActionInterface $action)
     {
         foreach ($products as $product) {
-            $this->productUpdater->setData(
+            $this->productFieldUpdater->setData(
                 $product,
                 $action->getField(),
                 $action->getValue(),