+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/{ → Repository}/PublishedAssociationRepository
 - Move PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/{ → Repository}/PublishedProductRepository
 - Add ProductBuilderInterface argument of the constructor of PimEnterprise/Bundle/CatalogRuleBundle/Validator/Constraints/ProductRule/ValueActionValidator
+- ProductsUpdater takes now ProductFieldUpdaterInterface as argument and not anymore ProductUpdaterInterface
+- ValueActionValidator takes now ProductFieldUpdaterInterface as argument and not anymore ProductUpdaterInterface
 
 # 1.3.x